pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "comply"
version = "0.1.0"
description = "A Python CLI tool and MCP server for enforcing team conventions via LLM"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
    "mcp[cli]>=1.0",
]

[project.optional-dependencies]
webhook = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "PyJWT[crypto]>=2.8",
]

[project.scripts]
comply = "comply.cli:app"
comply-mcp = "comply.mcp_server:serve"

[tool.hatch.build.targets.wheel]
packages = ["comply"]
